feat(examples): (WIP) `minimal`, `xihe`, and `depthanythingv2`
