+++ .vscode/settings.json
@@ -1,39 +1,35 @@
 {
-    "notebook.formatOnSave.enabled": true,
-    "notebook.codeActionsOnSave": {
-        "notebook.source.fixAll":"explicit",
-        "notebook.source.organizeImports": "explicit"
+  "notebook.formatOnSave.enabled": true,
+  "notebook.codeActionsOnSave": {
+    "notebook.source.fixAll": "explicit",
+    "notebook.source.organizeImports": "explicit"
+  },
+  "[python]": {
+    "editor.formatOnSave": true,
+    "editor.codeActionsOnSave": {
+      "source.fixAll": "explicit",
+      "source.organizeImports": "explicit"
     },
-    "[python]": {
-        "editor.formatOnSave": true,
-        "editor.codeActionsOnSave": {
-            "source.fixAll": "explicit",
-            "source.organizeImports": "explicit"
-        },
-        "editor.defaultFormatter": "charliermarsh.ruff",
-    },
-    "python.analysis.typeCheckingMode": "off", // TODO: Enable strict type checking
-    "python.analysis.autoImportCompletions": true,
-    "[csharp]": {
-        "editor.formatOnSave": true,
-        "editor.maxTokenizationLineLength": 2500,
-        "editor.inlineSuggest.suppressSuggestions": false
-    },
-    "cSpell.words": [
-        "arflow",
-        "astype",
-        "Behaviour",
-        "flipud",
-        "frombuffer",
-        "gmtime",
-        "ndarray",
-        "Protobuf",
-        "thecakelab",
-        "Xihe"
-    ],
-    "conventionalCommits.scopes": [
-        "server",
-        "client",
-        "others"
-    ],
+    "editor.defaultFormatter": "charliermarsh.ruff"
+  },
+  "python.analysis.typeCheckingMode": "off", // TODO: Enable strict type checking
+  "python.analysis.autoImportCompletions": true,
+  "[csharp]": {
+    "editor.formatOnSave": true,
+    "editor.maxTokenizationLineLength": 2500,
+    "editor.inlineSuggest.suppressSuggestions": false
+  },
+  "cSpell.words": [
+    "arflow",
+    "astype",
+    "Behaviour",
+    "flipud",
+    "frombuffer",
+    "gmtime",
+    "ndarray",
+    "Protobuf",
+    "thecakelab",
+    "Xihe"
+  ],
+  "conventionalCommits.scopes": ["server", "client", "examples", "others"]
 }
