--- python/examples/demo_xihe/scripts/launch_service.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-poetry run python examples/demo_xihe/xihe_server.py
