--- python/examples/demo_xihe/utils3d/__init__.py
@@ -1,7 +1,0 @@
-"""Basic utilities for 3D math operations
-"""
-from .container import *
-from .io import *
-from .math import *
-from .geometry import *
-from . import utils
