--- python/examples/demo_xihe/utils3d/container/__init__.py
@@ -1,4 +1,0 @@
-"""Tools for programming and managing 3D data
-"""
-from .vector import *
-from .matrix import *
