--- python/examples/demo_xihe/utils3d/geometry/__init__.py
@@ -1,3 +1,0 @@
-"""Tools for generating 3D geometries
-"""
-from .sphere import *
