--- python/examples/demo_xihe/utils3d/io/__init__.py
@@ -1,5 +1,0 @@
-"""Tools for IO operations
-"""
-from .image import *
-from .canvas import *
-from .point_cloud import *
