--- python/examples/demo_xihe/utils3d/math/__init__.py
@@ -1,5 +1,0 @@
-"""Tools for mathematical operations.
-"""
-from .transform import *
-from .coordinate import *
-from .spherical_harmonics import *
