+++ python/examples/minimal/pyproject.toml
@@ -1,0 +1,17 @@
+[tool.poetry]
+name = "minimal"
+version = "0.1.0"
+description = ""
+authors = ["Felix Nguyen <ngthinh302@gmail.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = ">=3.10, <3.13"
+arflow = { path = "../.." }
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+minimal = "minimal:main"
