+++ unity/Assets/Scripts/ARFlow/ARFlowClient.cs
@@ -60,7 +60,7 @@ namespace ARFlow
         /// Send a data of a frame to the server.
         /// </summary>
         /// <param name="frameData">Data of the frame. The typing of this is generated by Protobuf.</param>
-        public void SendFrame(DataFrameRequest frameData)
+        public string SendFrame(DataFrameRequest frameData)
         {
             string res = "";
             frameData.Uid = _sessionId;
@@ -81,6 +81,8 @@ namespace ARFlow
                 // Network, device image, camera intrinsics
                 Debug.LogError(e);
             }
+
+            return res;
         }
     }
 }
